Rename tags to metadata, add ARIA label
I think that metadata is a more sensible name for these things (in
terms of what the CSS class should be called). It gives more scope
to add other types of information in the same format.

Labelling them is slightly tricky because we don't do it for visual
users. There is an assumption that they are only meaningful if you
already know what a 'framework' or 'lot' are.

If we are going to label them then I think it makes more sense to
do so individually, rather than once for the whole list. That is
what I've done with this commit, but open to suggestions.