Update "multi select" with component code
What we have been calling "multi select" is now officially a GOV.UK component.
This commit replaces the code from Finder Frontend with the component version
from Static.

The point at which this is being brought in at is the merging of
[the pull request](https://github.com/alphagov/static/pull/556) I worked on
recently.

It will be a breaking change because of the markup changes.