Remove dependency on SSP naming of list items
The SSP expected the `name` attribute of list items to be in the format
'p1q5val1'.

Flask (and others) can provide an array of items if the request contains a
series of values with the same key, ie from a series of `input`s with the same
`name` attribute.

This commit removes the dependecy on naming convention in the SSP.