Table widths now 1/3 and 2/3 not 1/2 and 1/2
Having a narrower column on the left moves the content closer to the heading. It's easier to read as most of the headings are short (which means there's a unnecessarily large white space between content and heading). It also gives a more comfortable line length for the content column which is mostly the thing you want to spend more time reading and this is much better when the content has bullet points as they are the thing that suffers most if they get wrapped onto two line. On the downside there are some lines where the heading wraps and the content is short, this makes the row taller, but not any harder to read. In most cases it will make the overall page longer, but only by a small amount and on pages that are already long (so users are already scrolling).