Provide Jinja templates for use by apps
This commit does two main things:
- Add Jinja templates for all the design patterns in the toolkit, which can
  be used by any application that requires them
- Refactors the toolkit to generate the example markup from the Jinja templates
  and generate an example of the data passed to the template

This owes some debt to the work that @edds and @davidsingleton did with
GOV.UK components (http://govuk-component-guide.herokuapp.com/)

This commit also does some renaming of the patterns to reflect exactly what they
do. eg search-results (plural) is a template for rendering a list of search
results, whereas page-heading (singular) is a template for rendering a single
page heading. This commit only renames the examples and documentation, not the
actual toolkit files.