Change parameters of option select to match usage
This commit changes the parameters passed to option select to match how it is
used in the buyer app. This is a precursor to integrating the toolkit version
of option select into the buyer app.

The main change is specifying the `name` parameter for each checkbox
individually. This means that multiple 'questions' can now live inside one
`option-select`.

This commit also updates the documentation with an example of this usage.