Draw custom underlines for links in browse-list
We have noticed in user research and from feedback that support has received
that it is not always obvious that these are links.

Adding an underline to these links should help, but the default underlines
on large text are heavy and make the text less legible (or that's what we intuit
anyway).

The GOV.UK template uses a CSS border for drawing a nice, 1px underline on the
GOV.UK logo when hovered. So there is a visual precendent.

Using borders doesn't work so well on text, because they sit a lot lower than
the normal position of an underline. Using box-shadow instead allows control of
the vertical position of the underline, as well as the thickness (thickness is
managed by drawing two shadows, one for the blue and a white one to mask it).

For older browsers we fall back to a normal text-decoration: underline.