Use border not box-shadow to draw underlines
Using multiple box shadows caused problems in IE 11 because they did not line
up perfectly.

This commit removes the box shadows and instead applies a bottom border on the
link. It also adds a nested span inside the link in order to relatively position
the text at a vertical offset from the border.