Add summary.mapping_table macro for single-object tables
`summary.list_table` is convenient for lists of similar objects:
e.g. services, suppliers or drafts tables, But it's hard to use
for a table of service fields, where each field has a different type.

`mapping_table` creates a table structure, but doesn't take a
collection to iterate, so all table rows have to be created in the
parent template using the `summary.row` macro.

`summary.table` is renamed to `summary.list_table`, but the old
name is aliased for compatibility.