Add pricing form template
Add a template for the pricing form. The markup and SCSS is mostly taken
from the admin app [1]. This differs from other form templates in that
it is made up of multiple form elements and therefore requires multiple
default values. This is expected to be handled in macros in the
frontends. eg.

```jinja
{% macro pricing(question_content, service_data, errors) -%}
  {%
    with
    name=question_content.id,
    priceMin=service_data.priceMin,
    priceMax=service_data.priceMax,
    priceUnit=service_data.priceUnit,
    priceInterval=service_data.priceInterval,
    hint=question_content.hint,
    error=errors.get(question_content.id)['message']
  %}
    {% include "toolkit/forms/textbox.html" %}
  {% endwith %}
{%- endmacro %}
```

[1]
https://github.com/alphagov/digitalmarketplace-admin-frontend/commit/86afb5a119ed65c6f019ef9ab4649076a2562968