Don't show breadcrumb elements with missing labels
Only shows the breadcrumb element if it has a non-empty label value.
Since links and text with empty values are not visible anyway, this
hides the trailing ">" markers.

Both valid link and text breadcrumbs should have labels, but setting
label to an empty value is an easy way to filter the breadcrumbs on
some pages.