Link masthead to question wrapper, not field
This commit changes the link in the validation masthead to point at
the whole question, not just the field within it.

There is not always a one-to-one relationship between error and field,
eg where a question has a series of checkboxes. There is always a
one-to-one relationship between summary error message and question
wrapper.