Rebuild documentation when files are changed
This commit adds an extra script to run the Styleguide Publisher whenever files
in `./toolkit`, `./pages_builder/assets` or `./pages_builder/pages` are changed.

It also adds a block in the `Makefile` to run this script.

This should make the process of working on the toolkit locally a lot smoother,
because it removes the need to manually regenerate the documentation after a
change.