Add `multiquestion` macro
A multiquestion macro allows us to display multiple
answers in one summary table column cell.

To achieve this, it was necessary to separate the
`field()` calls (which create a new table cell)
from the markup containing the content.
This lets multiquestions call upon the content-only macros
without creating a whole whack of new table cells.

Also created a new file, `summary-content.html` with
the markup for presenting the service data, and retrofitted
the original macros to call upon the content-only versions.