Add bottom margin to temporary message
If there are multiple items in the page header (e.g. "DOS is coming" and "G7 is Closed" then the messages need a bit of space between them to look nicer. As the temporary message can potentially be followed by a range of possible elements it's better to add a bottom margin to te `temporary-message` rather than a top margin to all the other elements.