Rename "Back link" to "Secondary action link"
After talking to @tombye, it seems we've got a more general pattern than
"a link under a button that takes you back a page", which is
"a link under a button with takes you to a different page"

Basically, what it's a secondary action.

On a page where the primary action is "Save and continue", the secondary
action is "Return to summary".
On a page where the primary action is "Log in", the secondary action is
"Forgotten password".

Calling it a "Back link" conceptually limited what it might be used for,
so hopefully "Secondary action link" is a better name.