Abstract generic list classes
We have a number of list-based classes we want to be available to
all apps but aren't tied to a macro, so move them into a `_lists.scss`
file in a `/shared_scss` folder.
The fact that they are available to apps is just tacit knowledge at
this point, but probably @tombye will fix it.