Add spacing mixin to reduce ad-hoc styles
We're sometimes setting margins on lists by class, other times
by tag, and sometimes we're setting margin values meant to override
other ones in this file (like when we have a `list-bullet` inside of
a `check-list`)

Spacing macros bring a little bit of sanity to the list-spacing issue,
but not an enormous amount.
Better than before though, and @tombye had a look over it all.