Add markdown filter to advice and hints
When filling out briefs, most of the time we
have tons of `question_advice` which is all
going in as markdown.
It seems more reasonable to do the markdown
conversion here rather than just defining all
the CSS here but waiting for each app to do
markdown individually.

Also fixed a few classnames where underscores
were used instead of hyphens.