Bring in `.single-question-page` styles
We already have CSS that hides question labels if the
page title is going to be the same as the question label,
but previously it was only in the supplier app.

Since the buyer app is all about single question pages, pulling
them into the toolkit means that all apps have access to them.