Shrink negative right margin fixes everything
This fixes almost everything.

Having a negative right margin extends the length of the field
by percentage specificed.
If we had an element with 50px width and we applied a negative
right margin of -100%, our field will now take up 100px of width.

More to the point, by applying a negative margin of -100% to an
element in a 2/3 column, the element would stretch to 4/3 (four
thirds), breaking out of the grid altogether and, possibly,
extending the page horizontally.  This is what was happening to
our `.pricing` div.

Since the negative right margin is only applied to `.pricing`
divs inside of `.column-two-thirds` divs, we can safely extend
the right negative margin to -50% as the extra space will only
extend as far as the last third of the grid.

Moreover, applying the same negative margin to validation
wrappers was doubling the size of our already doubled `pricing`
div and we were ending up with an element with a width of had
8/3rds (ie, as though we had an `.eight-thirds-column` div).

In the case of the validation wrapper, we can just remove the
negative margin altogether as it doesn't seem to be doing
anything useful.