Bump up pricing breakpoint even further
So, it turns out that our breakpoint works perfectly until it
doesn't.

In our case, enforcing scroll bars on browser windows divorced the
page width from the browser width, which dropped the last field
down a line for very specific browser widths.

Bumping the breakpoint up 9 pixels to mitigate the problem as
much as possible without going over 960px (which is the max-width
we use for our central column).

No problems if scroll bars aren't present, and very unlikely to be
a problem otherwise.