Add `top` and `bottom` paragraphs to insruction list items
`top` and `bottom` give us more flexibility when building out
our instruction list sections.
Our two new fields allowed us to get rid of the `extra` field.
This is a breaking change, but we're only using the insruction list
in one app right now, so it should be easy to fix.