Support shown/hidden fields
Adds the ability for a boolean question to show or
hide 'text' or 'textarea' input fields.
[Docs for usage are here](https://github.com/alphagov/govuk_frontend_toolkit/blob/master/docs/javascript.md#showhide-content).

It's worth pointing out that the scope of this is
limited to these particular input types because
these are how we plan to do 'nice-to-have'
questions. It could be expanded to support other
fields if needed, but for now, these changes aren't
really general things we need to do.

Also worth pointing out that the `hidden` attribute
will create a `js-hidden` class which will only
truely hide the element if javascript is enabled.