Add advice option for link buttons
Link buttons were being rendered with a large margin on the top. This
was fine if there was no advice associated with the button, but meant
that the advice and the button looked ungrouped if there was.

By setting advice to true when using the template, it reduces the margin
to 15px and makes the advice and button look much more cohesive.