Fix issues with Sass compilation
The version of libsass-python needed bumping and
the asset compilation included the production of
unneeded files which prevented our use of
placeholders.

1. Bump version of libsass-python

Version 0.6.2 didn't play nicely with Macs running
El Capitan with the compilation of libsass
libsass-python does on install breaking due to OSX
moving c++ compilation to clang.

2. Remove step which produces CSS of each SCSS file

These files are not used and it is impossible to
generate some of them in isolation because they
depend on placeholders being present in a parent
stylesheet.