Replace markdown formatted strings with HTML
Within the self-contained universe of the frontend toolkit,
'markdown' as a concept has been obliterated.

It seemed like we had two options going forward:

1. Do some intermediary processing in the `generate_pages.py` file
   to run some of our question fields through markdown.
   This would reflects how these particular fields from the content
   will actually appear once rendered in front of users.
2. Remove all markdown from the examples in favour of straight HTML.
   This is consistent with how the templates itself act in isolation.

We've gone with the latter approach because in general it seems
safer to ignore the world outside of this app because it's less
certain and subject to change.