Remove underline from browse list headings
We wanted this when browse list was used for the Marketplace home page, but we don't any more.
Having headings inlined (which we needed for the fake underline to work)  is weird and could be prone to things looking bad.  So we're going to ditch the underline and leave the heading as block.