Create simple threaded server
To run the frontend toolkit thus far, we've been using a
`SimpleHTTPServer`, which is limited by the fact that it
can't serve simultaneous requests.

In our day-to-day usage, this hasn't been a huge problem, but
more recently we've tried running Wraith against it, which it
can't really handle. Even having a toolkit page open locally
in a browser while running Wraith would cause the server to
hang/fail.

To mitigate this, I created a super simple threaded version
of the `SimpleHTTPServer` based on a bunch of stackoverflow
answers that were all pretty much saying the same thing.

Usage is the same -- it just handles concurrent requests now.