Add gulp rule to copy govuk_toolkit stuff
Previously, updating the govuk_frontend_toolkit meant that the
stuff in the `node_modules` was updated but none of the code used
to build the frontend toolkit was. Copying the new stuff was left
as an exercise for the student.

Added a `copy` step to the gulpfile that copies all the stuff.