Update phase-banner styles
The govuk_frontend_toolkit has outlawed passing in the 'phase' to
the phase-banner components. Colour is now always the same, and it
is now just the text inside the phase tag that changes.

This represents a breaking change for our frontend apps.

Make alpha, beta and discovery colours $govuk-blue, in govuk_frontend_toolkit:
https://github.com/alphagov/govuk_frontend_toolkit/pull/370