Unify inner spacing of textareas and text fields
The actual text in our text fields was closer to the top
(and bottom) border than the text in our textareas.

Stopped setting an explicit height value on our text fields
and instead bumped the vertical padding. Means our
text fields look as before and our textareas are a bit
more spacious/luxurious.