Don't delete hidden files in `/pages` directory
By deleting the entire `/pages` directory, we broke the publishing
workflow.

Publishing our documentation looks like this:
- jenkins user starts on latest master branch
- remove and recreate empty `/pages` directory
- move into `/pages` directory and clone the frontend toolkit into it
- change branches to the `gh-pages` branch
- return to the root directory and delete contents of `/pages`
  (!except for hidden files!)
- from the root directory, generate the pages
  - all of the newly compiled HTML example pages are regenerated
    in the `/pages` directory
- return to the `/pages` directory, which is still its own repo
  and is still on the `gh-pages` branch
- add all the new pages and commit them
- push to the remote `gh-pages`

The mistake was adding a new gulp step as part of `generate_pages`
which deleted _everything_ in the /pages directory, including the
.git folder.  So when we added and commited everything, we would
be adding it to the master branch instead of the sub-repo in
pages.

Only deleting non-hidden files in the gulpfile fixes this.