Isolate selection button
The checkbox tree also needs to create radio and checkbox
`<input>`s but it doesn't want to always have to be creating a
list of them.
It's not always appropriate to have a `<fieldset>`, for example,
and we also never want to put `hint` or question_advice`
values into our nested lists of checkboxes.

Isolating the code that creates the `<input>`s keeps the
same output for selection buttons but lets other patterns
create checkboxes and radios as well.