Move `is_selected` logic into button partial
Previously, we were passing in an `is_selected` value
based on what our `input_value` said.

Unfortunately, within the `with` declaration, `input_value` hasn't
been created yet and so `is_selected` can never be true.

`is_selected` is also just a helper variable that we get
by checking the result of two others (`value` and `input_value`),
so we already pass in everything we need to assign it within
the macro itself.