Add more spacing to sticky element
Whether the element sticks or not, we want to maintain our
consistent vertical spacing rules.
However, if it _does_ stick to the top of the window, we want
a bit more of a cushion than our vertical spacing comes with.

Solution is to add a top-padding to the container element and then
a (nearly-)equal negative top-margin to pull it back to where
it was before.