Add default message to checkbox-tree floating message
Counter has been removed.
It is added by the JavaScript in the supplier app itself.
The sticky element stuff + keeping a running total of the number
of selected elements doesn't even show up without JavaScript, so
we don't need an empty counter element as a default, as that's not
helpful for anybody.