Check if another control is showing this content
If two controls are controlling one target, the last action wins.
It's possible to hide our target unintentionally.

For example, if we have two elements,
"A" and "B", both pointing at "C",
then we can hide "C" with "B",
even if "A" still says it should be revealed.

Like
A -> (show) -> C
B -> (hide) -> C

Then "C" will be hidden, even though we don't want it to be.
We only want it to be hidden if both "A" and "B" want to hide it.