Listen to change events rather than click events
This is interesting.

If you have a 'click' event handler on a radio button:

-> and then you do an in-browser click on the radio button:
     $(this).is(':checked') => true
     that is, inside the event handling function, the radio button
     that was clicked knows that it has been checked.

-> and then you call $('input').click() on the radio button:
     $(this).is(':checked') => false
     so, in this case, when inside the event handling function, the
     radio button doesn't know that it has been checked.

This seemed like suprising and unexpected behaviour to me.

There is another bit of logic somewhere that looks at the
`:checked` attribute of checkboxes to make a decision, so I
figured the behaviour would be uniform.

I found several old blog posts and a jQuery bug report about this
happening with checkboxes, but it appears it was only fixed for
checkboxes.  Radio buttons still can't consistently tell you if
they've been checked or not.

There seemed to be a couple of approaches to this:
- set a 1ms timeout on the body of the `deferred` method
- do some event introspection to figure out if it was a mouse click
  or a jquery click
- listen for `change` events instead

I've gone with the last one because it's the least disruptive and
all the tests still run green.

Sources:
- https://www.bennadel.com/blog/1525-jquery-s-event-triggering-order-of-default-behavior-and-triggerhandler.htm
- https://bugs.jquery.com/ticket/3827