Add last of type class
The :last-of-type pseudo selector isn't supported by ie8 so add last-of-type class in macro.  Last-of-type css class is being added within checkbox scss for now as this seems specific to this case rather than a general reset.

The new margin now means that when the focus is on a checkbox the yellow focus border extends beyond the grey connection border.  However when there is no focus the checkboxes are now aligned with grey border, matching service manual patterns.