New example page with lots of child questions
One of our problems around borders for child questions has been
that we just don't have a view of very many of them.

There are far more combinations of child/dependent questions in
the submissions process than we actually have in the frontend
toolkit in the 'combinations' page, yet that's where we normally
check when we're doing CSS changes.

Essentially, we're visually testing a very small subset of the
total univerise of possibilities and so we're going to get things
slip through.

This code creates a new page on which we can very rapidly create
loads of child fields one after another.

It's pretty prototype-y code (and the new page isn't included in
any of the menus), but it's helpful for seeing the effect of
general CSS changes to child questions.