Ignore `/toolkit` string from import paths
When linking to jinja2 partial templates, you have to give a local
path, which is appended to a template root path.
Our problem is that out template paths differ depending on whether
we're building pages in the toolkit itself or whether we're
building pages in our apps.

As an example, here is the path to `_selection-buttons.html`:

- buyer frontend: `app/templates/toolkit/forms/_selection-buttons.html`
- frontend-toolkit: `toolkit/templates/forms/_selection-buttons.html'

Since the `/toolkit` and `/templates` directories are reversed
across the repositories, it makes it hard for us to come up with
a common root path + path suffix.

For example if we want to import
`toolkit/forms/_selection-buttons.html`, that's fine in the buyer
app, but it doesn't work in the frontend toolkit because the
filepath is actually `templates/toolkit/forms/..`.

This commit doesn't give us general behaviour, but it does
give us a consistent way of importing things inside the toolkit
itself and the frontend apps.

For any include paths, the 'toolkit/' prefix will be removed.
So {% include 'toolkit/forms/textbox.html' %} will be understood
as {% include 'forms/textbox.html' %}, which fixes our specific
problem.

We could generalise this behaviour further, but at the moment we
don't need to.

Found this helpful: http://stackoverflow.com/a/3655911