Playing with refactoring the macros
The basic idea was this:

If we have

```
{% macro 1(a) %}
  {% with
      x,
      y,
      z,
      a
  %}
  {% endwith %}
{% endmacro %}
```

and

```
{% macro 2(b) %}
  {% with
      x,
      y,
      z,
      b
  %}
  {% endwith %}
{% endmacro %}
```

there are a bunch of repeat parameters in there.

It seemed like it would be possible to abstract some of the
variables into some kind of "base" macro and push be able to use
only the unique variables needed by the more specific macros.

(Un)fortunately, it seems like our tendency to use "includes"
statements everywhere precludes this kind of refactoring.

It's bad in the sense that we don't have more generic templates,
but probably it's also good that there's not much interpretation
work you need to do to see which variables exist and when they're
being passed through.