Update HTML for the keyword-search pattern
Noticed two things:

1. Our toolkit pattern is out of sync with the search input
   that's actually used in the buyer app
2. Our toolkit pattern is out of sync with the search input
   that's used on govuk search pages
   (ie, https://www.gov.uk/browse/driving)

As such, I've refactored the HTML to be closer to the one we
use in the buyer app, as well as made the input a type="search"
like the one GOV.UK is using.
Also used mostly the GOV.UK CSS rules for the search input.

HTML for the search input in the buyer app:
https://github.com/alphagov/digitalmarketplace-buyer-frontend/blob/edf3283e626b3e7eb6512382c1383bc34a787d43/app/templates/search/_services_filters.html#L26-L33

HTML for the search input in the GOV.UK header:
https://github.com/alphagov/govuk-headers-prototypes/blob/da56719fc8eb7826b9caa1dbf92467e6370a2c57/app/views/includes/search.html

CSS for the search input in the GOV.UK header:
https://github.com/alphagov/static/blob/bb2e0b052c15e5cfc58a0dca3a34bd4cc5caed46/app/assets/stylesheets/helpers/_header.scss#L102