Remove IE8 imcompatible CSS
Reverse margins and remove :last psuedo selectors.  Also refactor some of the CSS to make it simpler.  We should change the checkbox__tree classnames as this doesn't reflect the markup used in the table and is confusing, however this is outside the scope of the current pull request.  No visual changes.