Remove `visuallyhidden` class
Technically, both `visuallyhidden` and `visually-hidden` will work
as classnames to hide things in the visual interface (both are
included in govuk_template), but here in _our_ frontend toolkit,
we should be consistent.

`visually-hidden` is more readable and closer in convention to
our other classnames in, so I've removed the other ones.

I've also removed a `visually-hidden` CSS rule that was just a
straight copy from something we already get from govuk_template.