Capitalize the first letter of values in summary tables
In the frameworks content we store labels as all lowercase so we can use them in sentences (e.g. in the search summary sentence), but when displayed in the service page we want to see them with the first letter capitalised.

Python's `capitalize()` method is completely bonkers and makes the first letter capital but then lowercases everything else in the string it's applied to, so we need to only apply it to the first character and then append the rest.  I've tested to double-check and this won't break for one-character strings because calling `[1:]` on a one-character string gives an empty string, not an IndexError.