Markup changes and bring in show-hide content
We seem to be copying and pasting show-hide content rather than pulling it in at the moment which threw me.  The other javascript in this folder is quite custom so will need to look at this if we decide to just pull in show-hide content in the future.

Changed the markup to be inline with elements.