Add box-sizing mixin to options-container label
The labels for options in options containers have their width set to
100% of their parent. They also have a lot more padding on the left than
the right so that they don't clash with the checkbox.

Width doesn't take into account padding, which meant that the actual
label element was extending far too much to the right and so strings
that were long enough to need to be broken would disappear out of the
options container before breaking.

By setting the box-sizing to border-box, the padding is included and so
setting the width of the label to 100% prevents it from spilling out on
the right.

I've also adjusted the width for desktop's to 94% (from 90%). This
reduction in width is to allow for the scroll bar which is only present
when viewing on a desktop width screen. Due to the label now being less
wide that it previously was, leaving a 10% gap for the scroll bar looked
too much. Changing to 94% keeps it very similar to what it was.