Add a sass pattern and a macro for statistics
We added new patterns for statistics because no suitable ones
existed. This was done for the new feature that displays how many
suppliers applied for a brief on a brief opportunity page.
Trello ticket:
https://trello.com/c/9qlTFjK2/764-2-show-application-stats-on-opportunity-page

Turn big-stat.html into a macro

Make new examples for big_stat in page builder

Adjust big_stat macro for accessibility

Change file names for consistency and correct the version number

This was done in response to comments on a pull request:
https://github.com/alphagov/digitalmarketplace-frontend-toolkit/pull/363/files

Nest scss patterns for statistics followng code review