Remove reduced option select label width in desktop
For desktop, the width of the label was being reduced to allow room for
the scroll bar. At smaller widths under 796 pixels the scroll bar
dissapears and the labels become 100% width.

Unfortunatly if you have an option select in a narrow container on a
desktop page, then the desktop stylings are still applied and it still
has a weird gap and the scroll bar.

It makes sense to have the labels be 100% width at all times as on
chrome and safari space is left for the bar by default, and firefox the
label moves behind the scroll bar.