Add proposition-header template with Log in / Log out button
Extracts the common template for account / log in /log out links
header from the apps and adds it to the toolkit.

Changes "Log out" link to a form submit. Previously, log out action
was performed with a GET request, which means it wasn't protected
from CSRF attacks and a malicious page could log out the user.

Changing the request to POST requires changing the link to a form
with a submit button. None of our button styles fit this use case
and since "log out" has a pretty well defined behaviour we decided
to style it to match the link that it's replacing.