Add support for custom links in the proposition header
Allows replacing the admin-frontend propoistion header with the
shared template.

Admin app has additional links in the header based on the user role.
They link to other views within the app and only exist within the
admin app at the moment, so to avoid hardcoding the URLs we add a
way to pass them when including the template.