Make proposition-header use current_user attributes by default
Since frontend-toolkit templates are sometimes used outside the
Flask apps (eg when rendering documentation pages or when used in
prototypes) we can't rely on Flask or Flask-Login context objects.

There's no easy way to fake them in Jinja and Nunjucks, especially
when their use in templates involves method calls.

To avoid this, we replace them with separate variables that should
be set before including the template (eg with `with` or `set`).

But to make the use of templates in the apps more straightforward
we can use Flask objects as fallback values when the expected
variables aren't set.

For example, when using the proposition-header in the documentation
pages we set `logged_in` and `role` explicitly. But when used within
a Flask app the header can now be included without them.